pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pocketportal"
version = "4.6.0"
description = "One-for-all AI agent platform with modular architecture, stateful execution, lifecycle management, and enterprise features"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "PocketPortal Team"}
]
keywords = ["ai", "agent", "telegram", "chatbot", "llm", "ollama"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Core dependencies (always installed)
dependencies = [
    # Core Framework
    "python-telegram-bot==20.7",
    "python-dotenv==1.0.0",
    "pydantic==2.5.2",
    "pydantic-settings==2.1.0",
    "aiohttp==3.9.1",
    "aiofiles==23.2.1",
    "aiosqlite==0.19.0",
    "cryptography>=42.0.0",

    # Web Interface
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "websockets==12.0",

    # LLM Backends
    "ollama==0.1.6",

    # Configuration
    "pyyaml==6.0.1",
    "toml==0.10.2",

    # Logging
    "colorlog==6.8.0",
]

[project.optional-dependencies]
# Basic tool support
tools = [
    "qrcode[pil]==7.4.2",
    "pillow>=10.1.0",
    "requests==2.31.0",
    "beautifulsoup4==4.12.2",
    "lxml==4.9.4",
]

# Data processing tools
data = [
    "pandas==2.1.4",
    "numpy==1.26.2",
    "matplotlib==3.8.2",
    "py7zr==0.20.8",
]

# Document processing suite
documents = [
    "openpyxl==3.1.2",
    "python-docx==1.1.0",
    "python-pptx==0.6.23",
    "PyPDF2==3.0.1",
    "pandoc==2.3",
]

# Audio processing
audio = [
    "openai-whisper==20231117",
    "faster-whisper==0.10.0",
    "pydub==0.25.1",
]

# Knowledge and RAG
knowledge = [
    "sentence-transformers==2.2.2",
    "scikit-learn==1.3.2",
    "faiss-cpu==1.7.4",
]

# Automation and scheduling
automation = [
    "schedule==1.2.0",
    "apscheduler==3.10.4",
    "croniter==2.0.1",
]

# Browser automation
browser = [
    "playwright==1.40.0",
]

# Apple Silicon support (MLX)
mlx = [
    "mlx==0.0.9",
    "mlx-lm==0.0.9",
]

# Model Context Protocol
mcp = [
    "mcp==0.9.0",
]

# Observability and monitoring
observability = [
    "opentelemetry-api==1.21.0",
    "opentelemetry-sdk==1.21.0",
    "opentelemetry-exporter-otlp==1.21.0",
    "opentelemetry-instrumentation-fastapi==0.42b0",
    "prometheus-client==0.19.0",
]

# Distributed features (Redis for EventBroker, etc.)
distributed = [
    "redis==5.0.1",
    "hiredis==2.2.3",
]

# Development dependencies
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "black==23.12.0",
    "ruff==0.1.9",
    "mypy==1.7.1",
]

# Security enhancements (Docker sandbox)
security = [
    "docker==7.0.0",
]

# Complete installation with all features
all = [
    "pocketportal[tools,data,documents,audio,knowledge,automation,browser,mcp,security,observability,distributed]",
]

[project.urls]
Homepage = "https://github.com/ckindle-42/pocketportal"
Repository = "https://github.com/ckindle-42/pocketportal"
Issues = "https://github.com/ckindle-42/pocketportal/issues"

[project.scripts]
pocketportal = "pocketportal.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["pocketportal"]

[tool.setuptools.package-data]
pocketportal = ["prompts/*.txt", "web_static/*", "config/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Fast unit tests with no external dependencies (no I/O, no network, no database)",
    "integration: Integration tests requiring Docker, network, or database connections",
    "slow: Tests taking more than 5 seconds to complete",
    "requires_llm: Tests that require a running LLM backend",
    "requires_docker: Tests that require Docker to be running",
]
